+++ src/main/java/org/folio/services/domainevent/InstanceDomainEventPublisher.java
@@ -10,10 +10,13 @@ import io.vertx.core.Future;
 import java.util.Collection;
 import java.util.List;
 import java.util.Map;
+import org.apache.commons.collections4.CollectionUtils;
+import org.apache.commons.lang3.StringUtils;
 import org.apache.commons.lang3.tuple.Pair;
 import org.apache.logging.log4j.Logger;
 import org.folio.persist.InstanceRepository;
 import org.folio.rest.jaxrs.model.Instance;
+import org.folio.rest.jaxrs.model.PublishReindexRecords;
 
 public class InstanceDomainEventPublisher extends AbstractDomainEventPublisher<Instance, Instance> {
   private static final Logger log = getLogger(InstanceDomainEventPublisher.class);
@@ -22,6 +25,14 @@ public class InstanceDomainEventPublisher extends AbstractDomainEventPublisher<I
     super(new InstanceRepository(context, okapiHeaders),
       new CommonDomainEventPublisher<>(context, okapiHeaders,
         INSTANCE.fullTopicName(tenantId(okapiHeaders))));
+  }
+
+  public Future<Void> publishReindexInstances(String key, List<Instance> instances) {
+    if (CollectionUtils.isEmpty(instances) || StringUtils.isBlank(key)) {
+      return succeededFuture();
+    }
+
+    return domainEventService.publishReindexRecords(key, PublishReindexRecords.RecordType.INSTANCE, instances);
   }
 
   public Future<Void> publishInstancesCreated(List<Instance> instances) {