+++ src/test/java/org/folio/services/kafka/topic/KafkaAdminClientServiceTest.java
@@ -41,7 +41,9 @@ public class KafkaAdminClientServiceTest {
     "folio.foo-tenant.inventory.bound-with", "folio.foo-tenant.inventory.async-migration",
     "folio.foo-tenant.inventory.service-point", "folio.foo-tenant.inventory.classification-type",
     "folio.foo-tenant.inventory.location", "folio.foo-tenant.inventory.library",
-    "folio.foo-tenant.inventory.campus", "folio.foo-tenant.inventory.institution");
+    "folio.foo-tenant.inventory.campus", "folio.foo-tenant.inventory.subject-types",
+    "folio.foo-tenant.inventory.institution", "folio.foo-tenant.inventory.reindex-records",
+    "folio.foo-tenant.inventory.subject-sources");
   private KafkaAdminClient mockClient;
   private Vertx vertx;